requirements.txt
click==8.0.4
dataclasses==0.8
docopt==0.6.2
Flask==2.0.3
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
num2words==0.5.12
pycryptodome==3.15.0
pydantic==1.9.2
typing_extensions==4.1.1
Werkzeug==2.0.3
zipp==3.6.0
opcua~=0.98.13
numpy~=1.19.5
openpyxl~=3.0.10
pandas~=1.1.5
XlsxWriter~=3.0.3
websockets~=9.1
requests~=2.27.1
PyYAML~=6.0
pymongo~=4.1.1
tzlocal~=4.2
pytz~=2022.2.1
python-dateutil~=2.8.2
PyJWT~=1.7.1
Js2Py~=0.74
uModbus~=1.0.4